Update index.html
However, the file you provided has a different version number in the Firestore import. It should be consistent with the other Firebase imports.

3. ReferenceError: Can't find variable: loadingSpinner
This error means your code is trying to access a variable named loadingSpinner before it has been defined. This happens because the loadingSpinner variable is created inside your initApp() function, which only runs after your Firebase authentication code is finished.

Your code is trying to access loadingSpinner at the very beginning to hide it, but the initApp() function hasn't been called yet.

To fix this, you need to declare the loadingSpinner variable at the top level of your script, so it's available as soon as the page loads.

You should move the const loadingSpinner = document.getElementById('loading-spinner'); line outside of the initApp() function.